Create Question.txt
Please complete the tasks in the homework colab here

. When you're done, please go back to this page and click on `My Submissions` to upload your results. Please create a submission, upload `results.data`, then click `Submit`.

Your results.data file should include the predicted bounding boxes of your zombie detector, which will be compared against the reference boxes generated by a well trained model. Your grade will be computed based on the proportion of boxes that matched the reference boxes with a relative tolerance of 0.3. Given this, at least a 70% proportion is required for you to pass this assignment.

Good luck!